[backend] feat: ServiceRestaurant & Restaurant
Add function setRestaurantDescription() for model
Restaurant to update restaurant's description in
the database.

Add parameter restaurantDescription for function
insertRestaurant() in model Restaurant to follow
the change in the database.

Add parameter itemDescription for function
addItem() and editItem() in model Restaurant to
follow the change in the database.

Add parameter restaurantDescription for function
registerRestaurant() in class ServiceRestaurant to
follow the change in the database.

Configure function
getRestaurantPublicInformation() in class
ServiceRestaurant to return the restaurant
description.

Configure function setRestaurantInformation(), add
parameter 'description' and another if else
statement to allow restaurant owners to change
their restaurant description.

Modify test cases to incorporate new features.