[all] feat: Dynamic Image Generation Display
[backend]
Create path to display image in frontend from
backend in the folder 'images'.
Update functions getRestaurantByKeyword() and
getRestaurantByCategory() to send the image of
the restaurant.
Fix functions generateRestaurants() and
generateRestaurantItems() to assign images
correctly to generated data.

[frontend]:
Refactor the attribute 'src' of tag 'img' in
components CartItem, MenuItem, RestaurantCard,
RestaurantInfo to include the prefix:
`http://localhost:4000/` to correctly get the
image from backend.