[customer_interface] fix: Undefine Location
Fix error where the location, lat and lon in
Context will be turned into undefine after
navigating to different page like Cart or Profile.

The error is caused by the mechanism which allows
the home page to set the location, lat and lon
in the context based on what it retrieves from
the url, meaning that navigation from any page
other than landing page to home page will
have the location, lat and lon set to undefined
since they are not presented in the parameter.

The error is fixed by setting the location, lat
and lon on the context from the landing page
after an address is retrieved.

The location, lat and lon state in Context are
also grouped into a single state for better
state updating.